Fixed hanging test issue.
