+++ test/Microsoft.AspNetCore.SignalR.Tests/EndToEndTests.cs
@@ -2,11 +2,15 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System;
+using System.IO.Pipelines;
+using System.Net.Http;
 using System.Net.WebSockets;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Sockets.Client;
 using Microsoft.AspNetCore.Testing.xunit;
+using Microsoft.Extensions.Logging;
 using Xunit;
 
 namespace Microsoft.AspNetCore.SignalR.Tests
@@ -44,6 +48,43 @@ namespace Microsoft.AspNetCore.SignalR.Tests
 
                 await ws.CloseAsync(WebSocketCloseStatus.Empty, "", CancellationToken.None);
             }
+        }
+
+        [Fact]
+        public async Task CheckConnection()
+        {
+            const string message = "Major Key";
+            var baseUrl = _serverFixture.BaseUrl;
+            var loggerFactory = new LoggerFactory();
+
+            using (var httpClient = new HttpClient())
+            {
+                var transport = new LongPollingTransport(httpClient, loggerFactory);
+                using (var connection = await Connection.ConnectAsync(new Uri(baseUrl + "/echo"), transport, httpClient, loggerFactory))
+                {
+                    await connection.Output.WriteAsync(new Message(
+                        ReadableBuffer.Create(Encoding.UTF8.GetBytes(message)).Preserve(),
+                        Format.Text));
+
+                    var received = await ReceiveMessage(connection).OrTimeout();
+                    Assert.Equal(message, received);
+                }
+            }
+        }
+
+
+        private static async Task<string> ReceiveMessage(Connection connection)
+        {
+            Message message;
+            while (await connection.Input.WaitToReadAsync())
+            {
+                if (connection.Input.TryRead(out message))
+                {
+                    return Encoding.UTF8.GetString(message.Payload.Buffer.ToArray());
+                }
+            }
+
+            return null;
         }
     }
 }
